+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ### Bugfixes / Tweaks
 - Possible fix for #34: adding caps to all items causing incompatibility.
 - Added #19, do not display artifact bar when no artifacts equipped
+- Fix a performance issue with Built In Enchantments, when the enchantment is not found.